Metaprogress progress

+++ StortSpelprojekt/StortSpelprojekt/SaveState.h
@@ -12,7 +12,7 @@ struct SaveState
 	float playerFuel;
 	float playerFood;
 	bool showTutortial;
-	float upgradeCurrency;
+	unsigned int upgradeCurrency;
 	unsigned int highscore;
 	unsigned int nrOfGameWins;
 	static SaveState Default ()
@@ -24,7 +24,7 @@ struct SaveState
 		state.showTutortial = true;
 		state.segment = 0;
 		state.seed = rand();
-		state.upgradeCurrency = 0.f;
+		state.upgradeCurrency = 0;
 		state.highscore = 0;
 		state.nrOfGameWins = 0;
